Extract `print_yaml` function to utils
