+++ kpops/components/base_components/base_defaults_component.py
@@ -18,7 +18,7 @@ from kpops.utils.dict_ops import update_nested
 from kpops.utils.docstring import describe_attr
 from kpops.utils.environment import ENV
 from kpops.utils.pydantic import DescConfigModel, to_dash
-from kpops.utils.yaml_loading import load_yaml_file
+from kpops.utils.yaml import load_yaml_file
 
 try:
     from typing import Self
